requirements.txt
datasets>=2.14.0
huggingface-hub>=0.16.0
pandas>=2.0.0
numpy>=1.24.0
ipykernel
openai
google-generativeai
google-genai
python-dotenv